fix token gate to use ethereum for connection
