+++ lib/token-gates/index.ts
@@ -1,0 +1,1 @@
+export * from './token-gates';
