+++ pages/join.tsx
@@ -8,6 +8,7 @@ import getBaseLayout from 'components/common/BaseLayout/BaseLayout';
 import TokenGateForm from 'components/common/TokenGateForm';
 import Button from 'components/common/Button';
 import { useSpaces } from 'hooks/useSpaces';
+import { ErrorModal } from 'components/common/Modal';
 
 export function AlternateRouteButton ({ href, children }: { href: string, children: ReactNode }) {
   const [spaces] = useSpaces();
@@ -39,6 +40,7 @@ export default function CreateSpace () {
       router.push(`/${space.domain}`);
     }
   }
+
   useEffect(() => {
     if (spaces.some(space => space.domain === router.query.domain)) {
       router.push(`/${router.query.domain}`);
